Move the event handling logic to its own dedicated file
