+++ src/main.rs
@@ -3,7 +3,6 @@ use std::{io::Write, time::Duration};
 use clap::Parser;
 use crossterm::{
     cursor,
-    event::{KeyCode, KeyEvent, KeyEventKind, KeyModifiers},
     style::{style, Stylize},
     terminal, QueueableCommand,
 };
@@ -11,6 +10,7 @@ use crossterm::{
 mod ansi;
 mod config;
 mod entity;
+mod events;
 mod matrix;
 mod stream;
 mod symbols;
@@ -54,7 +54,7 @@ fn run(config: &config::Config) -> Result<(), Box<dyn std::error::Error>> {
         matrix.render(&config);
 
         // Handle events
-        if let Action::Exit = handle_events()? {
+        if let events::Action::Exit = events::handle_events()? {
             break;
         }
 
@@ -66,47 +66,6 @@ fn run(config: &config::Config) -> Result<(), Box<dyn std::error::Error>> {
     cleanup(&mut stdout)?;
 
     Ok(())
-}
-
-/// Instructs the main loop what to do
-enum Action {
-    /// Do nothing
-    None,
-    /// Exit the loop
-    Exit,
-}
-
-/// Processes and handles [crossterm events](crossterm::event). Returns an [`Action`] as a response.
-fn handle_events() -> std::io::Result<Action> {
-    if crossterm::event::poll(Duration::from_millis(100))? {
-        match crossterm::event::read()? {
-            crossterm::event::Event::Key(event) if event.kind == KeyEventKind::Press => {
-                return Ok(handle_key_event(event))
-            }
-            _ => (),
-        }
-    }
-    Ok(Action::None)
-}
-
-/// Handles keyboard events and returns an [`Action`] based on the key pressed.
-fn handle_key_event(event: KeyEvent) -> Action {
-    match event {
-        // Check if 'q', `Esc` or `Ctrl+C` has been pressed ...
-        KeyEvent {
-            code: KeyCode::Char('q'),
-            ..
-        }
-        | KeyEvent {
-            code: KeyCode::Esc, ..
-        }
-        | KeyEvent {
-            code: KeyCode::Char('c'),
-            modifiers: KeyModifiers::CONTROL,
-            ..
-        } => Action::Exit, // ... then respond with exit.
-        _ => Action::None, // ... otherwise, respond with none.
-    }
 }
 
 /// Prepares the terminal by switching to the alternate screen and clearing it.
